[layering] Moving BinarySearch and Sorting algorithms from math/mathcore/inc/TMath.h to TMathBase.h (evolution to ROOT "<algorithm>" header)
