+++ core/base/inc/TMathBase.h
@@ -30,6 +30,7 @@
 
 #include <cstdlib>
 #include <cmath>
+#include <algorithm>
 
 namespace TMath {
 
@@ -92,6 +93,17 @@ namespace TMath {
 
    //NextPrime is used by the Core classes.
    Long_t   NextPrime(Long_t x);   // Least prime number greater than x
+
+   // Binary search
+   template <typename T> Long64_t BinarySearch(Long64_t n, const T  *array, T value);
+   template <typename T> Long64_t BinarySearch(Long64_t n, const T **array, T value);
+   template <typename Iterator, typename Element> Iterator BinarySearch(Iterator first, Iterator last, Element value);
+
+   // Sorting
+   template <typename Element, typename Index>
+   void Sort(Index n, const Element* a, Index* index, Bool_t down=kTRUE);
+   template <typename Iterator, typename IndexIterator>
+   void SortItr(Iterator first, Iterator last, IndexIterator index, Bool_t down=kTRUE);
 }
 
 
@@ -244,5 +256,125 @@ inline ULong_t TMath::Range(ULong_t lb, ULong_t ub, ULong_t x)
 inline Double_t TMath::Range(Double_t lb, Double_t ub, Double_t x)
    { return x < lb ? lb : (x > ub ? ub : x); }
 
+template <typename Iterator, typename Element>
+Iterator TMath::BinarySearch(Iterator first, Iterator last, Element value)
+{
+   // Binary search in an array defined by its iterators.
+   //
+   // The values in the iterators range are supposed to be sorted
+   // prior to this call.  If match is found, function returns
+   // position of element.  If no match found, function gives nearest
+   // element smaller than value.
+
+   Iterator pind;
+   pind = std::lower_bound(first, last, value);
+   if ( (pind != last) && (*pind == value) )
+      return pind;
+   else
+      return ( pind - 1);
+}
+
+
+template <typename T> Long64_t TMath::BinarySearch(Long64_t n, const T  *array, T value)
+{
+   // Binary search in an array of n values to locate value.
+   //
+   // Array is supposed  to be sorted prior to this call.
+   // If match is found, function returns position of element.
+   // If no match found, function gives nearest element smaller than value.
+
+   const T* pind;
+   pind = std::lower_bound(array, array + n, value);
+   if ( (pind != array + n) && (*pind == value) )
+      return (pind - array);
+   else
+      return ( pind - array - 1);
+}
+
+template <typename T> Long64_t TMath::BinarySearch(Long64_t n, const T **array, T value)
+{
+   // Binary search in an array of n values to locate value.
+   //
+   // Array is supposed  to be sorted prior to this call.
+   // If match is found, function returns position of element.
+   // If no match found, function gives nearest element smaller than value.
+
+   const T* pind;
+   pind = std::lower_bound(*array, *array + n, value);
+   if ( (pind != *array + n) && (*pind == value) )
+      return (pind - *array);
+   else
+      return ( pind - *array - 1);
+}
+
+template<typename T>
+struct CompareDesc {
+
+   CompareDesc(T d) : fData(d) {}
+
+   template<typename Index>
+   bool operator()(Index i1, Index i2) {
+      return *(fData + i1) > *(fData + i2);
+   }
+
+   T fData;
+};
+
+template<typename T>
+struct CompareAsc {
+
+   CompareAsc(T d) : fData(d) {}
+
+   template<typename Index>
+   bool operator()(Index i1, Index i2) {
+      return *(fData + i1) < *(fData + i2);
+   }
+
+   T fData;
+};
+
+template <typename Iterator, typename IndexIterator>
+void TMath::SortItr(Iterator first, Iterator last, IndexIterator index, Bool_t down)
+{
+   // Sort the n1 elements of the Short_t array defined by its
+   // iterators.  In output the array index contains the indices of
+   // the sorted array.  If down is false sort in increasing order
+   // (default is decreasing order).
+
+   // NOTE that the array index must be created with a length bigger
+   // or equal than the main array before calling this function.
+
+   int i = 0;
+
+   IndexIterator cindex = index;
+   for ( Iterator cfirst = first; cfirst != last; ++cfirst )
+   {
+      *cindex = i++;
+      ++cindex;
+   }
+
+   if ( down )
+      std::sort(index, cindex, CompareDesc<Iterator>(first) );
+   else
+      std::sort(index, cindex, CompareAsc<Iterator>(first) );
+}
+
+template <typename Element, typename Index> void TMath::Sort(Index n, const Element* a, Index* index, Bool_t down)
+{
+   // Sort the n elements of the  array a of generic templated type Element.
+   // In output the array index of type Index contains the indices of the sorted array.
+   // If down is false sort in increasing order (default is decreasing order).
+
+   // NOTE that the array index must be created with a length >= n
+   // before calling this function.
+   // NOTE also that the size type for n must be the same type used for the index array
+   // (templated type Index)
+
+   for(Index i = 0; i < n; i++) { index[i] = i; }
+   if ( down )
+      std::sort(index, index + n, CompareDesc<const Element*>(a) );
+   else
+      std::sort(index, index + n, CompareAsc<const Element*>(a) );
+}
 
 #endif
